+++ examples/inventory/create_query.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
+read -p "Username: " CURL_USER
+
 INVENTORY_NAME="myInventory"
 
 echo "creating query"
-curl -f -H "Authorization: auBy4eDWrKWsyhiDp3AQiw" -H "Content-Type: text/plain" --data-binary @query.sql "http://localhost:8001/api/v1/inventory/$INVENTORY_NAME/query/myQuery"
+curl -f -u ${CURL_USER} -H "Content-Type: text/plain" --data-binary @query.sql "http://localhost:8001/api/v1/inventory/$INVENTORY_NAME/query/myQuery"