+++ examples/inventory/exec_query.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
+read -p "Username: " CURL_USER
+
 INVENTORY_NAME="myInventory"
 
 echo "exec query"
-curl -f -H "Authorization: auBy4eDWrKWsyhiDp3AQiw" -H "Content-Type: application/json" -d @query_params.json "http://localhost:8001/api/v1/inventory/$INVENTORY_NAME/query/myQuery/exec"
+curl -f -u ${CURL_USER} -H "Content-Type: application/json" -d @query_params.json "http://localhost:8001/api/v1/inventory/$INVENTORY_NAME/query/myQuery/exec"